// web-application/src/types-openapi/api/models/ListOrderRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListOrderRequest
 */
export interface ListOrderRequest {
    /**
     * 
     * @type {string}
     * @memberof ListOrderRequest
     */
    status?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListOrderRequest
     */
    paymentMethod?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ListOrderRequest
     */
    customerId?: number | null;
    /**
     * 
     * @type {Date}
     * @memberof ListOrderRequest
     */
    fromDate?: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof ListOrderRequest
     */
    toDate?: Date | null;
}

/**
 * Check if a given object implements the ListOrderRequest interface.
 */
export function instanceOfListOrderRequest(value: object): value is ListOrderRequest {
    return true;
}

export function ListOrderRequestFromJSON(json: any): ListOrderRequest {
    return ListOrderRequestFromJSONTyped(json, false);
}

export function ListOrderRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListOrderRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'] == null ? undefined : json['status'],
        'paymentMethod': json['paymentMethod'] == null ? undefined : json['paymentMethod'],
        'customerId': json['customerId'] == null ? undefined : json['customerId'],
        'fromDate': json['fromDate'] == null ? undefined : (new Date(json['fromDate'])),
        'toDate': json['toDate'] == null ? undefined : (new Date(json['toDate'])),
    };
}

export function ListOrderRequestToJSON(json: any): ListOrderRequest {
    return ListOrderRequestToJSONTyped(json, false);
}

export function ListOrderRequestToJSONTyped(value?: ListOrderRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
        'paymentMethod': value['paymentMethod'],
        'customerId': value['customerId'],
        'fromDate': value['fromDate'] === null ? null : ((value['fromDate'] as any)?.toISOString()),
        'toDate': value['toDate'] === null ? null : ((value['toDate'] as any)?.toISOString()),
    };
}