// web-application/src/types-openapi/api/models/ListOrderResponse.ts
export interface ListOrderResponse {
    /**
     * 
     * @type {Date}
     * @memberof ListOrderResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ListOrderResponse
     */
    updatedAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ListOrderResponse
     */
    createdBy?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListOrderResponse
     */
    updatedBy?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListOrderResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ListOrderResponse
     */
    orderCode?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListOrderResponse
     */
    bookingId?: string;
    /**
     * 
     * @type {OrderBookingCourtDto}
     * @memberof ListOrderResponse
     */
    booking?: OrderBookingCourtDto;
    /**
     * 
     * @type {string}
     * @memberof ListOrderResponse
     */
    bookingCourtOccurrenceId?: string;
    /**
     * 
     * @type {OrderBookingCourtOccurrenceDto}
     * @memberof ListOrderResponse
     */
    bookingCourtOccurrence?: OrderBookingCourtOccurrenceDto;
    /**
     * 
     * @type {number}
     * @memberof ListOrderResponse
     */
    customerId?: number;
    /**
     * 
     * @type {CustomerDto}
     * @memberof ListOrderResponse
     */
    customer?: CustomerDto;
    /**
     * 
     * @type {string}
     * @memberof ListOrderResponse
     */
    courtName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListOrderResponse
     */
    courtAreaName?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ListOrderResponse
     */
    courtTotalAmount?: number;
    /**
     * 
     * @type {number}
     * @memberof ListOrderResponse
     */
    courtPaidAmount?: number;
    /**
     * 
     * @type {number}
     * @memberof ListOrderResponse
     */
    courtRemainingAmount?: number;
    /**
     * 
     * @type {number}
     * @memberof ListOrderResponse
     */
    itemsSubtotal?: number;
    /**
     * 
     * @type {number}
     * @memberof ListOrderResponse
     */
    servicesSubtotal?: number;
    /**
     * 
     * @type {number}
     * @memberof ListOrderResponse
     */
    lateFeePercentage?: number;
    /**
     * 
     * @type {number}
     * @memberof ListOrderResponse
     */
    lateFeeAmount?: number;
    /**
     * 
     * @type {number}
     * @memberof ListOrderResponse
     */
    overdueMinutes?: number;
    /**
     * 
     * @type {number}
     * @memberof ListOrderResponse
     */
    totalAmount?: number;
    /**
     * 
     * @type {string}
     * @memberof ListOrderResponse
     */
    status?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListOrderResponse
     */
    paymentMethod?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListOrderResponse
     */
    note?: string | null;
    /**
     * 
     * @type {Array<PaymentDto>}
     * @memberof ListOrderResponse
     */
    payments?: Array<PaymentDto> | null;
    /**
     * 
     * @type {Array<OrderBookingOrderItemDto>}
     * @memberof ListOrderResponse
     */
    orderItems?: Array<OrderBookingOrderItemDto> | null;
    /**
     * 
     * @type {Array<OrderBookingServiceDto>}
     * @memberof ListOrderResponse
     */
    services?: Array<OrderBookingServiceDto> | null;
}